+++ CHANGES.md
@@ -370,4 +370,6 @@
 
 ### Pre-release (work-in-progress)
 - ReSound/Beltone/Interton/Danavox: update archived versions
+- ReSound/Beltone/Interton/Danavox: workaround for link error on GN's server
+- Quick Scan: remove unused scripts
 - Portable release: update Python modules